Added subtype/borrow
Passing a unique to borrow is different from unique/shared.
The latter doesn't require any encumbered permission for whole object.
The subtype/borrow rule handles this case.

Also changed signature of subtype again: now it takes a CM. The FS passed
in is not used in unique/shared case - in that case we'll lookup the CM
to find out the FM and FS for the class. I really don't like the definition
for the subtype right now, but haven't got better solution. :(

Also removed wf-fldty, and combined it with clsmap-fldmap.

Proved reftyping-ok/args for subtype/borrow case.