Finished method call checking (step 2)
Proved that if we have output/args converted from result/args, then
we can show that the method call can be proved.

Next is to prove that we can restore the permissions for the case
unique ==> borrowed. This is probably similar to let.elf.

This commit also replaces "declared-capset2perm" with "fldmap2perm".