sub done
For the sub rule, we need to change the handling of "fresh" type: added
a targets/fresh, which represents a fresh target after calling "new".
Also added an extra boolean to targets/unique, to indicate whether the
encumbered permission includes "fresh permissions". For this we also need
to add an extra existential variable of type permission (so now we have
3 existential variables in the output now).

For the sub rule, it's tricky to handle the case when you have two unique
targets. For instance, if the more restrict target is a borrowed case,
and the less restricted type is a unique+fresh target, then it's impossible
to derive the transformation, since how do you get a full permission for the
return value, plus the encumbered permissions? If the less restricted type
is a maybe null type, this is possible, but not for the non null case.

For simplicity, we simply disallow this kind of cases for now. For the two
unique targets case, we have two sub cases, targets`sub/bb and targets`sub/uu.
The former is for the cases where both sides are "borrowed" targets, and
the latter is for the cases where both sides are "unique" targets.

Reproved everything, now conversion.thm is missing the and case.