Proved 'and'
Changed the typing rule for 'and' a little bit - now for the second
conditional expression, the output should be the same as input. This is
because when converting this to permissions, we need to make sure that
both branches of the conditional in output should be the same. However,
if we have something like "a.f = new C" in the second cond expr, then
this may not be true anymore.