Proved 'if'
Need to prove quite a few lemmas. In particular, we need to show that
if we have two same 'results' and 'result2output' relation, then we can
derive that the two 'outputs' are also can be transformed to each other.
Here the output can be either expr or args. For cond, this cannot be proved,
since the formula in the output is unbound.