Fixes the special case for 'all' field
The 'all' field is special and should not be included when we
construct the class predicate. This changes the fldmap to not include
the field. For the type rule, when it's checking constructor, we add
the 'all' field to the domain of fldmap.

This also requires the fldmap to not contain 'all' field, e.g., the fldmap
must contain positive numbers only.

This seems like the easiest solution. Otherwise, if we keep the fldmap to be
the same, then we need to treat the 'all' field as a special case at every place
that we uses fldmap, for instance, make-encumbered, fldset2perm, etc.